+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Next
  - (Breaking) Changes the query hash in the webpack config to a file hash by default to prevent proxy caching.
+ - (Feature) Adds 'isMobile' computed to the '$viewport' helper/tools.
  - (Change) Changes remote asset protocols to https to prevent security exceptions on styleguide server.
  - (Change) Reverts router mode to 'history' for styleguide build.
  - (Change) Replaces relative imports with alias based ones.